rename: Module 네이밍 변경, 디렉토리 이동
