+++ app/src/main/java/woowacourse/shopping/ui/common/RepositoryModule.kt
@@ -1,11 +1,11 @@
-package woowacourse.shopping.ui
+package woowacourse.shopping.ui.common
 
 import woowacourse.shopping.data.CartRepository
 import woowacourse.shopping.data.CartRepositoryImpl
 import woowacourse.shopping.data.ProductRepository
 import woowacourse.shopping.data.ProductRepositoryImpl
 
-object RepositoryModules {
+object RepositoryModule {
     val productRepository: ProductRepository = ProductRepositoryImpl()
     val cartRepository: CartRepository = CartRepositoryImpl()
 }
