+++ app/src/main/java/woowacourse/shopping/ui/common/ViewModelFactory.kt
@@ -2,7 +2,6 @@ package woowacourse.shopping.ui.common
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import woowacourse.shopping.ui.RepositoryModules
 
 class ViewModelFactory : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
@@ -12,7 +11,7 @@ class ViewModelFactory : ViewModelProvider.Factory {
         val types = constructor.parameterTypes
         val params = mutableListOf<Any?>()
 
-        val applicationProperties = RepositoryModules::class.java.declaredFields
+        val applicationProperties = RepositoryModule::class.java.declaredFields
         for (type in types) {
             val field = applicationProperties.first { it.type == type }
                 ?: throw IllegalArgumentException("Can't find Property $type")
